Really fix future kernel compatibility
Earlier commit made sure we don't error out if the kernel capability
version is unknown; this ensures compatibility with future kernels.

Looking at the code, I realized p.hdr.version should be initialized to
linuxCapVer3 in that case, not the version returned by the kernel,
otherwise we supply v3 data structure with (say) v4 version set in
header.

Practically, this was not a real bug (yet) because v4 is not (yet)
available, but if it will ever be introduced later, this fix makes us
ready.

Fixes: 558410569e2151 ("Fix future version compatibility")
Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>